+++ SpatialGDK/Source/SpatialGDKEditor/Public/SpatialGDKEditorSettings.h
@@ -121,7 +121,6 @@ struct FWorkerTypeLaunchSection
 		, bAutoNumEditorInstances(true)
 		, NumEditorInstances(1)
 		, bManualWorkerConnectionOnly(false)
-		, WorkerLoadBalancing(nullptr)
 	{
 	}
 
@@ -160,9 +159,6 @@ struct FWorkerTypeLaunchSection
 	/** Determines if the worker instance is launched manually or by SpatialOS. */
 	UPROPERTY(Category = "SpatialGDK", EditAnywhere, config, meta = (DisplayName = "Manual worker connection only"))
 	bool bManualWorkerConnectionOnly;
-
-	UPROPERTY(Transient, Category = "SpatialGDK", EditAnywhere, Instanced)
-	UAbstractRuntimeLoadBalancingStrategy* WorkerLoadBalancing;
 };
 
 USTRUCT()
@@ -177,11 +173,8 @@ struct FSpatialLaunchConfigDescription
 		FWorkerTypeLaunchSection UnrealWorkerDefaultSetting;
 		UnrealWorkerDefaultSetting.bManualWorkerConnectionOnly = true;
 
-		ServerWorkersMap.Add(SpatialConstants::DefaultServerWorkerType, UnrealWorkerDefaultSetting);
+		ServerWorkerConfig = UnrealWorkerDefaultSetting;
 	}
-
-	/** Set WorkerTypesToLaunch in level editor play settings. */
-	SPATIALGDKEDITOR_API void OnWorkerTypesChanged();
 
 	/** Deployment template. */
 	UPROPERTY(Category = "SpatialGDK", EditAnywhere, config)
@@ -196,7 +189,7 @@ struct FSpatialLaunchConfigDescription
 	TArray<FWorkerTypeLaunchSection> ServerWorkers_DEPRECATED;
 
 	UPROPERTY(Category = "SpatialGDK", EditAnywhere, EditFixedSize, config)
-	TMap<FName, FWorkerTypeLaunchSection> ServerWorkersMap;
+	FWorkerTypeLaunchSection ServerWorkerConfig;
 };
 
 /**
@@ -236,16 +229,7 @@ public:
 	virtual void PostEditChangeProperty(struct FPropertyChangedEvent& PropertyChangedEvent) override;
 	virtual void PostInitProperties() override;
 
-	void OnWorkerTypesChanged();
-
-private:
-
-	/** Set DAT in runtime settings. */
-	void SetRuntimeUseDevelopmentAuthenticationFlow();
-	void SetRuntimeDevelopmentDeploymentToConnect();
-
 public:
-
 	/** If checked, show the Spatial service button on the GDK toolbar which can be used to turn the Spatial service on and off. */
 	UPROPERTY(EditAnywhere, config, Category = "General", meta = (DisplayName = "Show Spatial service button"))
 	bool bShowSpatialServiceButton;
@@ -304,7 +288,7 @@ private:
 	UPROPERTY(EditAnywhere, config, Category = "Snapshots", meta = (DisplayName = "Snapshot to load"))
 	FString SpatialOSSnapshotToLoad;
 
-	UPROPERTY(EditAnywhere, config, Category = "Schema Generation", meta = (Tooltip = "Platform to target when using Cook And Generate Schema"))
+	UPROPERTY(EditAnywhere, config, Category = "Schema Generation", meta = (Tooltip = "Platform to target when using Cook And Generate Schema (if empty, defaults to Editor's platform)"))
 	FString CookAndGeneratePlatform;
 
 	UPROPERTY(EditAnywhere, config, Category = "Schema Generation", meta = (Tooltip = "Additional arguments passed to Cook And Generate Schema"))
@@ -372,10 +356,6 @@ public:
 	UPROPERTY(EditAnywhere, config, Category = "Assembly")
 	FString BuildSimulatedPlayerExtraArgs;
 
-	/** If the Development Authentication Flow is used, the client will try to connect to the cloud rather than local deployment. */
-	UPROPERTY(EditAnywhere, config, Category = "Cloud Connection")
-	bool bUseDevelopmentAuthenticationFlow;
-
 	/** The token created using 'spatial project auth dev-auth-token' */
 	UPROPERTY(EditAnywhere, config, Category = "Cloud Connection")
 	FString DevelopmentAuthenticationToken;
@@ -404,11 +384,17 @@ private:
 	static bool IsManualWorkerConnectionSet(const FString& LaunchConfigPath, TArray<FString>& OutWorkersManuallyLaunched);
 
 public:
-	UPROPERTY(EditAnywhere, config, Category = "Mobile", meta = (DisplayName = "Connect to a local deployment"))
-	bool bMobileConnectToLocalDeployment;
+	/** If checked, use the connection flow override below instead of the one selected in the editor when building the command line for mobile. */
+	UPROPERTY(EditAnywhere, config, Category = "Mobile", meta = (DisplayName = "Override Mobile Connection Flow (only for Push settings to device)"))
+	bool bMobileOverrideConnectionFlow;
 
-	UPROPERTY(EditAnywhere, config, Category = "Mobile", meta = (EditCondition = "bMobileConnectToLocalDeployment", DisplayName = "Runtime IP to local deployment"))
-	FString MobileRuntimeIP;
+	/** The connection flow that should be used when pushing command line to the mobile device. */
+	UPROPERTY(EditAnywhere, config, Category = "Mobile", meta = (EditCondition = "bMobileOverrideConnectionFlow", DisplayName = "Mobile Connection Flow"))
+	TEnumAsByte<ESpatialOSNetFlow::Type> MobileConnectionFlow;
+
+	/** If specified, use this IP instead of 'Exposed local runtime IP address' when building the command line to push to the mobile device. */
+	UPROPERTY(EditAnywhere, config, Category = "Mobile", meta = (DisplayName = "Local Runtime IP Override"))
+	FString MobileRuntimeIPOverride;
 
 	UPROPERTY(EditAnywhere, config, Category = "Mobile", meta = (DisplayName = "Mobile Client Worker Type"))
 	FString MobileWorkerType = SpatialConstants::DefaultClientWorkerType.ToString();
@@ -425,8 +411,9 @@ public:
 	UPROPERTY(EditAnywhere, config, Category = "Launch", meta = (EditCondition = "bGenerateDefaultLaunchConfig", DisplayName = "Launch configuration file options"))
 	FSpatialLaunchConfigDescription LaunchConfigDesc;
 
-	UPROPERTY(EditAnywhere, config, Category = "SpatialGDK")
-	TEnumAsByte<ESpatialOSNetFlow::Type> SpatialOSNetFlowType = ESpatialOSNetFlow::LocalDeployment;
+	/** Select the connection flow that should be used when starting the game with Spatial networking enabled. */
+	UPROPERTY(EditAnywhere, config, Category = "Connection Flow", meta = (DisplayName = "SpatialOS Connection Flow Type"))
+	TEnumAsByte<ESpatialOSNetFlow::Type> SpatialOSNetFlowType;
 
 	FORCEINLINE FString GetSpatialOSLaunchConfig() const
 	{
@@ -452,10 +439,7 @@ public:
 			: SpatialOSSnapshotToLoad;
 	}
 
-	FORCEINLINE FString GetCookAndGenerateSchemaTargetPlatform() const
-	{
-		return CookAndGeneratePlatform;
-	}
+	FString GetCookAndGenerateSchemaTargetPlatform() const;
 
 	FORCEINLINE FString GetCookAndGenerateSchemaAdditionalArgs() const
 	{
@@ -639,7 +623,12 @@ public:
 
 	bool IsDeploymentConfigurationValid() const;
 
-	void SetRuntimeDevelopmentAuthenticationToken();
+	void SetDevelopmentAuthenticationToken(const FString& Token);
+	void SetDevelopmentDeploymentToConnect(const FString& Deployment);
+
+	void SetExposedRuntimeIP(const FString& RuntimeIP);
+
+	void SetSpatialOSNetFlowType(ESpatialOSNetFlow::Type NetFlowType);
 
 	static bool IsProjectNameValid(const FString& Name);
 	static bool IsAssemblyNameValid(const FString& Name);